Multiple Plastic Detection
Code Detects multiple plastics within the region of interests. I then detects the edges of the plastic and calculates the area of the plastic according to a 2d image. It takes measurment of the area into a CSV file and takes a screenshot of the whole image and a screen shot of the ROI at different time intervals