Broken plastic test
Plastic that had an area of ~4.10, that was broken in 2 uneven pieces and then the total area was measured from the two pieces. This was done to see if the code could still detect the plastic and it's according area. It was found that it could. however the area was slightly different. This can be because of the position of the plastic compared to the camera or that the the camera now also sees the edges that are sticking out of the plastic.